Fix Muxlicer bug when getting repeated first step, added Hexmix VCA tooltips/bypasses.
Fixes #32. Couple of things going on - one is that the 1ms grace period on reset doesn't really work in this context, and breaks EOC chaining which is a common use case. Two, if the clock is already high when a reset was requested, then the boolean trigger detector doesn't detect a rising edge.